requirements.txt
azure-identity
azure-core
langchain
langchain-cohere
requests
cohere
langchain-community
faiss-cpu